+++ pyoxidizer/src/starlark/python_executable.rs
@@ -400,48 +400,15 @@ impl PythonExecutable {
         Ok(Value::new(NoneType::None))
     }
 
-    /// PythonExecutable.add_in_memory_package_resource(resource)
-    pub fn starlark_add_in_memory_package_resource(
+    /// PythonExecutable.add_python_package_resource(resource, location=None)
+    pub fn starlark_add_python_package_resource(
         &mut self,
         type_values: &TypeValues,
         resource: &Value,
+        location: &Value,
     ) -> ValueResult {
         required_type_arg("resource", "PythonPackageResource", &resource)?;
-
-        let raw_context = get_context(type_values)?;
-        let context = raw_context
-            .downcast_ref::<EnvironmentContext>()
-            .ok_or(ValueError::IncorrectParameterType)?;
-
-        let r = match resource.downcast_ref::<PythonPackageResource>() {
-            Some(r) => Ok(r.inner.clone()),
-            None => Err(ValueError::IncorrectParameterType),
-        }?;
-        info!(
-            &context.logger,
-            "adding in-memory resource data {}",
-            r.symbolic_name()
-        );
-        self.exe
-            .add_python_package_resource(&r, Some(ConcreteResourceLocation::InMemory))
-            .map_err(|e| {
-                ValueError::from(RuntimeError {
-                    code: "PYOXIDIZER_BUILD",
-                    message: e.to_string(),
-                    label: "add_in_memory_package_resource".to_string(),
-                })
-            })?;
-
-        Ok(Value::new(NoneType::None))
-    }
-
-    /// PythonExecutable.add_package_resource(resource)
-    pub fn starlark_add_package_resource(
-        &mut self,
-        type_values: &TypeValues,
-        resource: &Value,
-    ) -> ValueResult {
-        required_type_arg("resource", "PythonPackageResource", &resource)?;
+        let location = OptionalResourceLocation::try_from(location)?;
 
         let raw_context = get_context(type_values)?;
         let context = raw_context
@@ -458,49 +425,12 @@ impl PythonExecutable {
             r.symbolic_name()
         );
         self.exe
-            .add_python_package_resource(&r, None)
+            .add_python_package_resource(&r, location.into())
             .map_err(|e| {
                 ValueError::from(RuntimeError {
                     code: "PYOXIDIZER_BUILD",
                     message: e.to_string(),
-                    label: "add_package_resource".to_string(),
-                })
-            })?;
-
-        Ok(Value::new(NoneType::None))
-    }
-
-    /// PythonExecutable.add_filesystem_relative_package_resource(prefix, resource)
-    pub fn starlark_add_filesystem_relative_package_resource(
-        &mut self,
-        type_values: &TypeValues,
-        prefix: &Value,
-        resource: &Value,
-    ) -> ValueResult {
-        let prefix = required_str_arg("prefix", &prefix)?;
-        required_type_arg("resource", "PythonPackageResource", &resource)?;
-
-        let raw_context = get_context(type_values)?;
-        let context = raw_context
-            .downcast_ref::<EnvironmentContext>()
-            .ok_or(ValueError::IncorrectParameterType)?;
-
-        let r = match resource.downcast_ref::<PythonPackageResource>() {
-            Some(r) => Ok(r.inner.clone()),
-            None => Err(ValueError::IncorrectParameterType),
-        }?;
-        info!(
-            &context.logger,
-            "adding executable relative resource data {}",
-            r.symbolic_name()
-        );
-        self.exe
-            .add_python_package_resource(&r, Some(ConcreteResourceLocation::RelativePath(prefix)))
-            .map_err(|e| {
-                ValueError::from(RuntimeError {
-                    code: "PYOXIDIZER_BUILD",
-                    message: e.to_string(),
-                    label: "add_filesystem_relative_package_resource".to_string(),
+                    label: "add_python_package_resource".to_string(),
                 })
             })?;
 
@@ -753,9 +683,11 @@ impl PythonExecutable {
                 optimize_level,
                 &Value::from("in-memory"),
             ),
-            "PythonPackageResource" => {
-                self.starlark_add_in_memory_package_resource(type_values, resource)
-            }
+            "PythonPackageResource" => self.starlark_add_python_package_resource(
+                type_values,
+                resource,
+                &Value::from("in-memory"),
+            ),
             "PythonPackageDistributionResource" => {
                 self.starlark_add_package_distribution_resource(type_values, resource)
             }
@@ -809,10 +741,10 @@ impl PythonExecutable {
                 optimize_level,
                 &Value::from(format!("filesystem-relative:{}", prefix)),
             ),
-            "PythonPackageResource" => self.starlark_add_filesystem_relative_package_resource(
+            "PythonPackageResource" => self.starlark_add_python_package_resource(
                 type_values,
-                prefix,
                 resource,
+                &Value::from(format!("filesystem-relative:{}", prefix)),
             ),
             "PythonPackageDistributionResource" => self
                 .starlark_add_filesystem_relative_package_distribution_resource(
@@ -868,7 +800,11 @@ impl PythonExecutable {
                 optimize_level,
                 &Value::from(NoneType::None),
             ),
-            "PythonPackageResource" => self.starlark_add_package_resource(type_values, resource),
+            "PythonPackageResource" => self.starlark_add_python_package_resource(
+                type_values,
+                resource,
+                &Value::from(NoneType::None),
+            ),
             "PythonPackageDistributionResource" => {
                 self.starlark_add_package_distribution_resource(type_values, resource)
             }
@@ -1091,25 +1027,9 @@ starlark_module! { python_executable_env =>
     }
 
     #[allow(non_snake_case, clippy::ptr_arg)]
-    PythonExecutable.add_in_memory_package_resource(env env, this, resource) {
+    PythonExecutable.add_python_package_resource(env env, this, resource, location=NoneType::None) {
         match this.clone().downcast_mut::<PythonExecutable>()? {
-            Some(mut exe) => exe.starlark_add_in_memory_package_resource(&env, &resource),
-            None => Err(ValueError::IncorrectParameterType),
-        }
-    }
-
-    #[allow(non_snake_case, clippy::ptr_arg)]
-    PythonExecutable.add_filesystem_relative_package_resource(env env, this, prefix, resource) {
-        match this.clone().downcast_mut::<PythonExecutable>()? {
-            Some(mut exe) => exe.starlark_add_filesystem_relative_package_resource(&env, &prefix, &resource),
-            None => Err(ValueError::IncorrectParameterType),
-        }
-    }
-
-    #[allow(non_snake_case, clippy::ptr_arg)]
-    PythonExecutable.add_package_resource(env env, this, resource) {
-        match this.clone().downcast_mut::<PythonExecutable>()? {
-            Some(mut exe) => exe.starlark_add_package_resource(&env, &resource),
+            Some(mut exe) => exe.starlark_add_python_package_resource(&env, &resource, &location),
             None => Err(ValueError::IncorrectParameterType),
         }
     }